// src/interfaces.ts
export interface addHouseState {
  home_id: number;
  home_name: string;
  home_address: string;
  home_price: number;
  home_img: string;
  home_disc: string;
}
export interface displayHouseState {
  home_name: string;
  home_address: string;
  home_price: number;
  home_img: string;
  home_disc: string;
}